cleaning and splitting
